# services/audit/api/openapi.yaml
openapi: 3.1.0
info:
  title: GUVA Audit Service
  version: 0.1.0
  description: |
    Append-only, hash-chained audit ledger for every action that affects the security
    or integrity of the GUVA platform.

    **Reads only.** Writes happen via Kafka — producers emit `audit.entry.appended.v1`
    events on `ug.go.guva.audit.entry.appended.v1`, and this service is the sole writer
    to the chain. See [pkg/platform/audit](../../pkg/platform/audit) for the producer
    library that every service uses.

    ## Chain invariant

    For every entry `n`: `entry_hash = SHA256(canonical_row || previous_hash)`.
    `previous_hash` of entry 1 is the literal 64-zero string (genesis). Modification of
    any past entry breaks every subsequent hash. Append-only is also enforced at the
    DB layer: `BEFORE UPDATE OR DELETE` triggers `RAISE EXCEPTION`.

    ## Authentication

    Bearer JWT, gated by the `audit:read` scope.
servers:
  - url: http://localhost:8000/v1/audit
    description: Local — via APISIX gateway (recommended)
  - url: http://localhost:7072
    description: Local — direct to the service (bypasses gateway; dev only)
  - url: https://api.guva.go.ug/v1/audit
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /entries:
    get:
      tags: [read]
      summary: List audit entries (cursor pagination + filters)
      parameters:
        - in: query
          name: after
          description: Entry id cursor; returns rows with entry_id > after.
          schema: { type: integer, format: int64, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
        - in: query
          name: actor_id
          schema: { type: string }
        - in: query
          name: subject_id
          schema: { type: string }
        - in: query
          name: action
          description: |
            Event type. Producer-defined; current vocabulary includes
            `identity.consumer.created`, `audit.entries.queried`,
            `audit.chain.verified` (the last two are meta-audit events
            emitted by this service when its own read endpoints are called).
          schema: { type: string, example: "identity.consumer.created" }
        - in: query
          name: result
          schema: { type: string, enum: [ok, denied, error, inconclusive] }
        - in: query
          name: from
          description: Inclusive lower bound on occurred_at (RFC 3339).
          schema: { type: string, format: date-time }
        - in: query
          name: to
          description: Exclusive upper bound on occurred_at (RFC 3339).
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Page of entries
          content:
            application/json:
              schema:
                type: object
                required: [entries, count, next_cursor]
                properties:
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/Entry' }
                  count: { type: integer }
                  next_cursor:
                    type: integer
                    format: int64
                    description: Pass as `after` on the next request; 0 means no more rows.
              examples:
                default:
                  value:
                    count: 1
                    next_cursor: 5
                    entries:
                      - entry_id: 5
                        entry_uuid: "0ad6b85b-1971-0c9e-aaaa-bbbbbbbbbbbb"
                        occurred_at: "2026-05-24T18:00:00.000Z"
                        actor_kind: "service"
                        actor_id: "identity"
                        subject_kind: "consumer"
                        subject_id: "acacia-onboarding"
                        action: "identity.consumer.created"
                        result: "ok"
                        correlation_id: "9369c308-d96d-4905-9096-f369161cdb77"
                        detail: { agency_name: "Acacia", consumer_id: "..." }
                        previous_hash: "0000000000000000000000000000000000000000000000000000000000000000"
                        entry_hash:    "0ad6b85b19710c9e..."
        "401": { description: missing or invalid bearer token }
        "403": { description: token does not carry audit:read }
        "400": { description: malformed filter }
        "500":
          description: |
            Either the read itself failed, or the meta-audit emit failed
            and the response was withheld so the read does not leak
            untraced. `type` distinguishes the two: `internal_error` vs
            `meta_audit_failed`.
  /verify:
    get:
      tags: [integrity]
      summary: Walk the chain and verify hash integrity
      description: |
        Recomputes the hash for every entry in the range and checks the
        previous_hash linkage. Use to prove that a snapshot has not been
        tampered with. The walk stops at the first break.
      parameters:
        - in: query
          name: from_id
          schema: { type: integer, format: int64, default: 1 }
        - in: query
          name: to_id
          description: Pass 0 (or omit) to walk to the end of the chain.
          schema: { type: integer, format: int64, default: 0 }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      ok:      { type: boolean, enum: [true] }
                      from_id: { type: integer, format: int64 }
                      to_id:   { type: integer, format: int64 }
                  - type: object
                    properties:
                      ok:          { type: boolean, enum: [false] }
                      broken_at:   { type: integer, format: int64 }
                      broken_uuid: { type: string }
                      detail:      { type: string }
              examples:
                ok:
                  value: { ok: true, from_id: 1, to_id: 100 }
                broken:
                  value: { ok: false, broken_at: 42, broken_uuid: "abc...", detail: "previous_hash mismatch" }
        "401": { description: missing or invalid bearer token }
        "403": { description: token does not carry audit:read }
        "500":
          description: |
            Either the verify walk failed, or the meta-audit emit failed
            and the response was withheld. `type` distinguishes the two:
            `internal_error` vs `meta_audit_failed`.
  /export:
    get:
      tags: [export]
      summary: Sign and return a chain range as a verifiable bundle
      description: |
        Returns a signed JSON bundle containing the requested chain
        range plus the anchor (entry_hash of the row preceding the
        range, or genesis). Verifiers use `pkg/platform/audit` —
        either as a Go import or via the standalone `audit-verify`
        binary in `tools/audit-verify/` — to:

          1. Verify the Ed25519 signature against canonical JSON.
          2. Walk the chain inside the bundle, recomputing each
             entry_hash and asserting the chain link from `anchor`.

        Bundles cap at 500 entries; for larger ranges, page by
        adjusting from_id. Each call also emits an
        `audit.bundle.exported` meta-audit event.
      parameters:
        - in: query
          name: from_id
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
        - in: query
          name: to_id
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: A signed bundle
          content:
            application/json:
              schema:
                type: object
                required:
                  - format_version
                  - generator
                  - generated_at
                  - range_from_id
                  - range_to_id
                  - anchor
                  - entries
                  - signing_pubkey
                  - signature
                properties:
                  format_version: { type: string, example: "1" }
                  generator: { type: string, example: "guva-audit" }
                  generated_at: { type: string, format: date-time }
                  range_from_id: { type: integer, format: int64 }
                  range_to_id: { type: integer, format: int64 }
                  anchor:
                    type: object
                    properties:
                      anchor_entry_id: { type: integer, format: int64 }
                      anchor_entry_hash: { type: string, description: "hex SHA-256; 64×'0' for genesis" }
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/Entry' }
                  signing_pubkey:
                    type: string
                    description: Base64 (std) of the Ed25519 public key the signature was created with.
                  signature:
                    type: string
                    description: Base64 (std) of the Ed25519 signature over canonical JSON of the bundle with signature blanked.
        "400": { description: malformed from_id / to_id }
        "401": { description: missing or invalid bearer token }
        "403": { description: token does not carry audit:read }
        "500":
          description: |
            Internal failure (list, sign, or meta-audit emit). The bundle
            is withheld if meta-audit fails — same fail-closed policy
            as /entries and /verify.
...
  /anchors:
    get:
      tags: [anchors]
      summary: List Merkle anchors over the chain (most recent first)
      description: |
        Each anchor is a binary-Merkle root over a contiguous range of
        chain entries. The audit service computes one every
        `AUDIT_ANCHOR_INTERVAL` (default 5m). An operator may publish
        a root to an external witness; the receipt is stored in
        `external_proof`. See docs/AUDIT.md §"External Merkle anchoring".
      parameters:
        - in: query
          name: after
          description: Cursor — return anchors with anchor_id < after.
          schema: { type: integer, format: int64 }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        "200":
          description: Page of anchors
          content:
            application/json:
              schema:
                type: object
                properties:
                  anchors:
                    type: array
                    items: { $ref: '#/components/schemas/Anchor' }
                  count: { type: integer }
                  next_cursor: { type: integer, format: int64 }
        "401": { description: missing or invalid bearer token }
        "403": { description: token does not carry audit:read }
  /anchors/{id}:
    get:
      tags: [anchors]
      summary: Get a single anchor by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: The anchor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Anchor' }
        "404": { description: no anchor with that id }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Anchor:
      type: object
      required: [anchor_id, range_from_id, range_to_id, leaf_count, merkle_root, algorithm, computed_at]
      properties:
        anchor_id:      { type: integer, format: int64 }
        range_from_id:  { type: integer, format: int64 }
        range_to_id:    { type: integer, format: int64 }
        leaf_count:     { type: integer, format: int64 }
        merkle_root:    { type: string, description: "hex SHA-256" }
        algorithm:      { type: string, example: "sha256-binary-merkle-v1" }
        computed_at:    { type: string, format: date-time }
        external_proof:
          description: |
            Operator-attached witness receipt (Sigstore Rekor entry,
            blockchain tx hash, consortium-ledger pointer, etc.).
            Absent until the operator publishes the root externally.
          oneOf:
            - type: object
            - type: 'null'
    Entry:
      type: object
      required: [entry_id, entry_uuid, occurred_at, actor_kind, actor_id,
                 action, result, previous_hash, entry_hash]
      properties:
        entry_id:       { type: integer, format: int64 }
        entry_uuid:     { type: string, format: uuid }
        occurred_at:    { type: string, format: date-time }
        actor_kind:     { type: string, example: "service" }
        actor_id:       { type: string, example: "identity" }
        subject_kind:   { type: string }
        subject_id:     { type: string }
        action:         { type: string, example: "identity.consumer.created" }
        result:         { type: string, enum: [ok, denied, error, inconclusive] }
        correlation_id: { type: string, format: uuid }
        detail:
          description: Service-defined; small, stable JSON.
          oneOf:
            - type: object
            - type: array
            - type: string
            - type: number
            - type: boolean
            - type: 'null'
        previous_hash:  { type: string, description: "Hex SHA-256; 64 zeros for the first entry" }
        entry_hash:     { type: string, description: "Hex SHA-256 of canonical(row || previous_hash)" }